backend/requirements.txt
django
djangorestframework
pinecone
langchain
langchain-google-vertexai
langchain-pinecone
beautifulsoup4
requests
tqdm
python-dotenv
google-cloud-speech
google-cloud-texttospeech 